Refactor bibs pagination into BibList
